Fix Jangaroo 3 inter-module dependencies.
Sometimes, mavenProjectModel.getName() returns null. Then, the plugin fell
back to using mavenProjectModel.getFinalName() for the Flex build configuration
name.
But getFinalName() includes the version number, which is inconsistent with the way
Flex build configuration are references. Thus, the non-versioned references resulted
in dead links.
Now, instead, the plugin uses getMavenId().getArtifactId(). This works, but is still
not fool-proof, because there could be two modules with the same artifactId, but
different groupId, but since this is a rare case that can be avoided for Jangaroo
modules, I went for the shorter and thus more readable build config names.