IDEA-27 Jangaroo 0.9 plugin does not work with IDEA 13
Works again, but make/build is not yet supported, because
we have to implement external build (JPS) that was introduced
with IDEA 12 and remains the only option with IDEA 13.
Also, lost a minor EXML feature: due to changes in IDEA's
internal XML Psi API, it is no longer possible to navigate from
an EXML element to an ActionScript class if the other EXML
file is not available. Navigation will then lead to the XSD element
as usual. This is not so bad, as all EXML elements should be
defined in EXML by now (no more hand-written AS config classes!)
and all documentation is in the XSD, so documentation lookup
still works.

Note that this commit contains changes that are *not compatible
with IDEA 12*. We now have a branch "idea-12" to update the
IDEA-12-compatible version of the plugin.