IDEA-27 Jangaroo 0.9 plugin does not work with IDEA 13
Gave up serializing Jangaroo SDK's JAR paths as "additional data".
Instead, using the Jangaroo SDK home path to compute JAR locations
in the JPS plugin. To achieve this, moved JAR path computation from
JangarooSdkType to JpsJangarooSdkType and used File instead of VirtualFile,
so the code can be used in JPS.
To access the JPS SDK, changed JangarooFacetImporter to set the
Jangaroo SDK as module SDK instead of as a JangarooFacet property.
TODO: get rid of Jangaroo SDK chooser in facet.