IDEA-27 Jangaroo 0.9 plugin does not work with IDEA 13
Fixed that generated AS files were not picked up by AS->JS compiler.
Seems although we already tell the IDEA build process that we just
generated a source file, it also has to be marked dirty explicitly to
trigger the next compiler.
Also simplified filesToBeCompiled data structure: all dirty files from
the current moduleBuildTarget must be from the same module, and
the forTests flag also must be the same for one moduleBuildTarget.