Back to old behavior where to set the Jangaroo SDK. Temporarily, we used the module SDK to transport the Jangaroo SDK and its home path to JPS. However, this no longer allows mixing Java and Jangaroo code within one module (which is not a good idea, but used to be possible), and does not fit the concept of an IDEA Facet that should add an aspect to a module, not re-interpret a module setting. We could have made a custom Jangaroo module type, but if Jangaroo settings are in a Facet, all of them should be in there. That said, I changed the Jangaroo JPS plugin to retrieve the JpsSdk from the name stored in the Jangaroo Facet (it was not so easy to find out how to retrieve a JpsSdk by name, but finally managed to do so using only public IDEA API). Setting the module SDK in JangarooFacetImporter is no longer necessary and has been removed. So now, the overall behavior on Maven-importing and building is very similar to how it was in IDEA 12.
On the fly, fixed that the JPS plugin tried to compile all ActionScript and EXML code, even when
there is now Jangaroo-/EXML-Facet. The reason was that JangarooModelSerializerExtension returned
an empty configuration if none was found, so that every module looked like it had a facet.
Now, existence of the corresponding Facet is checked first, and files from modules that do not have
the corresponding Facet are silently skipped. Also, an error is reported if an EXML Facet comes
without a Jangaroo Facet.