Add some logging to Jangaroo Builder.
Each start and end of a module compilation is logged as "info".
For each module, compiler classpath, sourcepath and sourcefiles are
 logged as "debug" (change build-log.properties to see this).
If no Jangaroo build configuration can be found for a Jangaroo
module, this is logged as "warn".
If the Jangaroo compiler should return an unknown compilation
result code, this is logged as "error" (incompatible compiler?).
(cherry picked from commit 8e4e5b9)