Remove usages of "gone" Ext JS API.
If you map a class member to empty string, all usages will be removed.
In MXML, attributes are deleted (cannot comment inside attributes!),
in ActionScript, code using the member is wrapped in a block comment
with the prefix EXT6_GONE: followed by the fully-qualified name of the
member that is gone.