Download Ext AS 6 and optimize imports in migration tool
It's no longer necessary to add a Maven dependency to Ext AS 6 in
some pom of the project. The migration tool does not expect it to be
available as project library but downloads it via Maven itself.

After usage search and before the actual migration, the classes of the
project's Ext AS library are changed to point to Ext AS 6 so that new
classes and members from the migration map can be resolved.
After migration the project library is restored to point to the
original classes (of Ext 3.4).

As part of the migration imports and XML namespace declarations get
optimized for all touched files. This is necessary because
imports are messed up (e.g. duplicate imports) by the actual
migration. An additional manual "Optimize Imports" step is no longer
needed.

The migration map is no longer loaded via GlobalSearchScope index
lookup but directly from the downloaded Ext AS jars via JarFileSystem.
A search scope index would only work with actually declared and
used project libraries but we don't add Ext AS 6 as separate project
library in Idea's LibraryTable anymore.

The build version of the old Ext AS library is no longer hard-coded.
Any 2.0.x version is recognized. (The build version of the new Ext AS
 library is still hard-coded)