Fix IDEA-59: Cannot run compiler: FileNotFoundException: commons-lang3-3.2.1.jar
https://jangaroo.myjetbrains.com/youtrack/issue/IDEA-59

Revamp Jangaroo SDK Maven import to use
jangaroo-compiler-x.y.z-jar-with-dependencies.jar.
On Maven import, existing Jangaroo SDKs with the target version are
updated or a new SDK is created, now containing only this one JAR.
Since this JAR is not downloaded during the Maven build, the IDEA
plugin now has to care about this and does so by using the IDEA API
MavenRunner with the goal dependency:get.

On the fly, removed some backwards-compatibility code that used
reflection to work for two incompatible versions of some IDEA API.