Fix IDEA-64 Downloading Jangaroo Compiler during Maven Import Fails
https://jangaroo.myjetbrains.com/youtrack/issue/IDEA-64

As workaround for https://youtrack.jetbrains.com/issue/IDEA-232152, we no
longer use MavenRunner.runBatch(), but the underlying method
MavenRunConfigurationType.runConfiguration() directly.
Since this method is async and uses a callback, we now keep track of all
downloading Jangaroo compiler artifacts to ensure that only one request
is triggered for each Jangaroo SDK at any given time. The Jangaroo SDK
is created synchronously, but setupSdkPaths() is called later, when the
compiler artifact has been downloaded successfully.

An on-the-fly fix is that the JRE to use for the Maven call is now set
explicitly to the first available Java JRE, since otherwise, the project
SDK is used, which may happen to be not set at all or be set to another
SDK type (like a Jangaroo SDK), not a JRE.