Really Fix IDEA-64 Downloading Jangaroo Compiler during Maven Import Fails
https://jangaroo.myjetbrains.com/youtrack/issue/IDEA-64

Fix of the fix:
When using new MavenRunnerSettings(), the default is to use the global project
SDK to run Maven.
Since especially in a Jangaroo project, the global project SDK may not even be
a Java JDK, this default is a risky choice for running Maven.
So to avoid this, I used IDEA's API
    ProjectJdkTable.getInstance().getAllJdks()
foolishly assuming it would return only Java JDKs, and simply using the first
on ([0]). In local tests, we were lucky and the first "JDK" actually was a
JDK, but as it turned out after releasing 4.2018.20, the API above returns
all *SDKs*, making its name's choice something that can only be explained by
"historic reasons". At the end of the day, my attempt to make the JDK choice
more stable ended up in making it always wrong for some users (namely those
that have a non-JDK as first entry in their "JDK" table).
The obvious fix is to use the JDK the developer chose in her "Maven > Runner"
settings. After some digging around in IDEA community source, I found out
that these can be obtained by
    MavenRunner.getInstance(project).getSettings()
Since I want to complement these settings, they need to be cloned.
Then, no more need to set a custom JRE.

On the fly, improved error message when Jangaroo SDK cannot be downloaded.
Without deleting Jangaroo SDKs first, a Maven (Re)Import will not help.