More detailed error logging when failing to find JPS build configuration
It happens that the Maven import process creates Flex modules,
but in the *.iml file, the <module> root tag only has a version,
but no type="Flex" attribute. In IDEA, such modules are still
shows as Flex modules (small "F" at their folder icon), but in
JPS, they deserialize as Java modules :-( This can be fixed by
Maven-reimporting, and if this does not help, add the type="Flex"
attribute manually.
Let's at least add detailed logging if the requested module is
indeed found, but is not a JPS Flex module, so that we can
analyze that this error happened in the build log.